Phishing-Domain-Detection-main/requirements.txt
flask
gunicorn
scipy
scikit-learn
numpy
pandas
python-whois
urllib3
requests
bs4
ipaddress
lxml
